Add files via upload
I made this webpage to learn about CSS and different ways of styling through CSS. The website is clean, simple, and not to mention beautiful! Apart from hyperlinks, there isn't much content and the paragraphs and sentences are mostly sample text. This is because the focus of this project was to focus on the aesthetics and styling.